+++ src/Stack/Options/CleanParser.hs
@@ -2,21 +2,23 @@
 module Stack.Options.CleanParser where
 
 import           Options.Applicative
-import           Stack.Clean                       (CleanOpts (..))
+import           Stack.Clean                       (CleanCommand(..), CleanOpts (..))
 import           Stack.Prelude
 import           Stack.Types.PackageName
 
 -- | Command-line parser for the clean command.
-cleanOptsParser :: Parser CleanOpts
-cleanOptsParser = CleanShallow <$> packages <|> doFullClean
+cleanOptsParser :: CleanCommand -> Parser CleanOpts
+cleanOptsParser Clean = CleanShallow <$> packages <|> doFullClean
   where
     packages =
         many
             (packageNameArgument
                  (metavar "PACKAGE" <>
-                  help "If none specified, clean all local packages"))
+                  help "If none specified, clean all project packages"))
     doFullClean =
         flag'
             CleanFull
             (long "full" <>
-             help "Delete all work directories (.stack-work by default) in the project")
+             help "Delete the project’s stack working directories (.stack-work by default).")
+
+cleanOptsParser Purge = pure CleanFull