+++ test/integration/tests/3863-purge-command/files/src/Lib.hs
@@ -1,0 +1,4 @@
+module Lib where
+
+someFunc :: ()
+someFunc = ()