+++ ChangeLog.md
@@ -60,6 +60,8 @@ Behavior changes:
 * When using `stack script`, custom snapshot files will be resolved
   relative to the directory containing the script.
 
+* Remove the deprecated `--upgrade-cabal` flag to `stack setup`.
+
 Other enhancements:
 
 * Defer loading up of files for local packages. This allows us to get
@@ -116,6 +118,8 @@ Other enhancements:
 * `stack clean` will delete the entire `.stack-work/dist` directory,
   not just the relevant subdirectory for the current GHC version. See
   [#4480](https://github.com/commercialhaskell/stack/issues/4480).
+* Add `stack purge` as a shortcut for `stack clean --full`. See
+  [#3863](https://github.com/commercialhaskell/stack/issues/3863).
 
 Bug fixes: